Added more plugin to nvim  and improvements

+++ install.sh
@@ -83,8 +83,16 @@ vim +NeoBundleInstall +qall +UpdateRemotePlugins
 
 # Creating symbolic link for nvmrc
 mkdir -p ~/.config/nvim/
-ln -s ~/.vimrc ~/.config/nvim/init.vim
-ln -s ~/.vim ~/.nvim
+
+if [ -e "~/.vimrc" ]; then
+  rm -f ~/.vimrc
+  ln -s ~/.vimrc ~/.config/nvim/init.vim
+fi
+
+if [ -e "~/.vim" ]; then
+  rm -f ~/.vim
+  ln -s ~/.vim ~/.nvim
+fi
 
 # Install tmux plugin manager
 if [ ! -e ~/.tmux/plugins/tpm ]; then
@@ -103,7 +111,14 @@ else
 fi
 
 #Adding Fonts to nvim devicons
-cd ~/Library/Fonts && curl -fLo "Droid Sans Mono for Powerline Nerd Font Complete.otf" https://raw.githubusercontent.com/ryanoasis/nerd-fonts/master/patched-fonts/DroidSansMono/complete/Droid%20Sans%20Mono%20for%20Powerline%20Nerd%20Font%20Complete.otf
+echo "Adding DroidSansMono font for webdev icons.."
+curl -fLo "Droid Sans Mono for Powerline Nerd Font Complete.otf" https://raw.githubusercontent.com/ryanoasis/nerd-fonts/master/patched-fonts/DroidSansMono/complete/Droid%20Sans%20Mono%20for%20Powerline%20Nerd%20Font%20Complete.otf
+
+cp ./Droid\ Sans\ Mono\ for\ Powerline\ Nerd\ Font\ Complete.otf ~/Library/Fonts
 
 echo "vim/bundle/youcompleteme/install.py --gocode-completer --tern-completer"
 #~/.vim/bundle/youcompleteme/install.py --gocode-completer --tern-completer
+
+# Call +UpdateRemotePlugins
+nvim +UpdateRemotePlugins
+
